Add direct links to the current state of the program on GitHub at the end of a session.
Add symbolic names for every step in the repository so that the links stay valid when the commits change. Unsure whether to use lightweight tags or branches for that. Given the non-replacement behavior of changed tags discussed in `git help tag`, branches might make it easier on followers. Let's see.